Currently working on adding some basic specifier support to circumvent basic <format> usage. Still haven't implemented a compile time parser for the message pattern to take advantage of <format>'s implementations if an unsupported type or spec is found.
Also still need to map the SpecValues over to GetArgValue() and implement
the means of formatting the arguments based on that SpecValue.

A WHOLE lot of work needs to be done still in order to make this work,
but on the upside, the current parsing and storing of spec types and spec values
made a negligible hit on performance (~50 to ~60 nanoseconds) so there's some hope I guess.

The current downside is that if a spec actually is present, for the most part, the
non-compile-time check on unsupported specifiers and the resultant call to <format>
hits performance hard