Lots of small updates and changes.
Changed a lot of functions to take in the string destination and eliminate
a bunch of string allocations, removed SpecValue enum and vector. I wasn't
using it and due to forgetting about it, never cleared it causing massive
memory usage to occur.

Sped up alignment functions, added a more sane way to parse precision with
fill alignment as well as nested brackets. Added a method to keep track of
how many valid brackets are in the message to format in order to only parse
the message and format an argument for how many brackets were found.

I plan on possibly adding a container of start/end
point pairs to reduce bracket searching since it's already done in when
accounting for how many valid brackets are available; this way, when parsing
the message, I can just directly get the bracket to parse and format via these
point pairs.