This commit has a lot of changes: - Changed some function paramaters - Added a ReserveCapacity() function as well as a few others - Overhauled a few functions' logic - Removed the IsValid'x'Spec() functions as well as some others - Other small changes here and there
These changes did speed up the flat substitution timings, however, they
were not the focus at this moment; the focus being on complex formatting.
With that being said, for the complex format string currently in Sandbox.cpp,
the timings so far are ~0.47us for se_format_to(), ~0.53us for se_format(), and
~2.33us for std::vformat_to().

I still need to focus on the localization functions
as they are still, by themselves, very slow (or at least slower than they need to be).
Afterwards, I need to add a custom handler as well as add tests for that and then I
think I'll be finished over on this side and can move my attention back over to the
logger side of things.

The reason for the shift in focus on performance of argument formatting was because now
that MSVC back-ported their fixes to <format>, I needed to create a reason to use the native
argument formatting functions (se_format() and se_format_to()) after putting in the time I did
into writing an implementation of these two. I believe I've accomplished that since these functions
are now roughly anywhere between 20%-50% faster with a huge gain in performance; for instance,
formatting a string argument leads to ~1800%-~1900% performance increase, although for a single char,
it's for some reason a ~17% decrease in performance but IS still within 0.5% of MSVC's implementation,
even in that worst case. All other formatting, is much faster in every case. Localization can be made
more efficient I believe, hence the note on working on that soon. The only benefit right now that MSVC's
implementation has is the type erasure of std::vformat family and the custom formatting handlers.
(At least, as far as I know right now).