Lots of changes!
So the last commit was definitely just lucky for the laptop; on my desktop
with an intel i9-10900 and on the older server with an AMD ryzen 3600, last
commit was actually quite a bit slower. In lieu of that, decided to revert
changes back to the older method.

On top of that, moved everything that was possible to specify as 'constexpr'over to
ArgFormatterImpl.h; locale specific functions and variables prevented constexpr-fying
everything, including the constructor. Since making the constructor constexpr in and
of itself made the complex format pattern currently in Sandbox.cpp ~11ns faster, I
decided to remove the locale reference variable and the SetLocale() function. Also
separated localized and non-localized formatting due to this change and now I understand
why one of the possible reasons libfmt and <format> had overloads for this was due to
exactly the non-constexpr-ness of locale. I kind of doubt it, but it would be pretty
cool if it were possible to use or work-around the non-constness of the locale specific
functions in order to work in a constexpr context. If I can manage that, then this
ArgFormatter deal will just be a drop in header-only implementation

The goal of the shift is to hopefully add some static or constexpr structures in order to
get some compile time warnings similar to what libfmt and <format>(post back-ported fixes)
already have. It would also be really awesome to make some of the constexpr functions with
return types (like FindBrackets() for instance) have consteval overloads so that if a formatting
function only contains const arguments, then that argument is formatted at compile time instead.
Obviously would need to utilize something like 'is_constant_evaluated()' for any runtime or
non-const arguments though.
- Note: I still need to add custom formatting handlers in order to be fully compatible.